do not delete groups that were specified from open id groups w/ burnettk

+++ spiffworkflow-backend/src/spiffworkflow_backend/services/authorization_service.py
@@ -506,7 +506,9 @@ class AuthorizationService:
                 )
             else:
                 for desired_group_identifier in desired_group_identifiers:
-                    UserService.add_user_to_group_by_group_identifier(user_model, desired_group_identifier)
+                    UserService.add_user_to_group_by_group_identifier(
+                        user_model, desired_group_identifier, source_is_open_id=True
+                    )
                 current_group_identifiers = [g.identifier for g in user_model.groups]
                 groups_to_remove_from_user = [item for item in current_group_identifiers if item not in desired_group_identifiers]
                 for gtrfu in groups_to_remove_from_user:
@@ -947,7 +949,8 @@ class AuthorizationService:
         added_group_identifiers.add(SPIFF_GUEST_GROUP)
         groups_to_delete = GroupModel.query.filter(GroupModel.identifier.not_in(added_group_identifiers)).all()  # type: ignore
         for gtd in groups_to_delete:
-            db.session.delete(gtd)
+            if not gtd.source_is_open_id:
+                db.session.delete(gtd)
 
         for wugam in initial_waiting_group_assignments:
             if wugam not in added_waiting_group_assignments:
